fix: generare validation help script using projen
